Rewrite keyboard navigation for accessibility
Rewrite of the keyboard control functions to better match 'normal' form behaviour. Tab to enter the control and cursor to move around within the control while enter and spacebar calls the click function. When you leave the control either by keyboard or mouse it places a tabstop to either the selected value if one exits or the first available item.